Adds support for PETSc vectors.
This patch introduces a new class in the petsc namespace: Vector. Wraps around
Vec from petsc and provides operations and functions to manipulate vector.

The design is slightly radical, in the sense that it aims to be safe and
function-oriented, much like petsc. Iterators are not provided, as they provide
several opportunity for data races, and raw memory access is forbidden.
However, several features are provided to operate on the vector as a vector,
which should cover needs.

The design is currently quite minimal - as we discover needs this should be
extended with more features and functions. Methods should only deal with the
vector's structure, memory and lifetime. Features that operate and compute
should be implemented as operators or free functions.