Adds support for PETSc matrices.
This patch introduces the Matrix class into the Petsc namespace. Wraps around
Mat and provides methods and functions to manipulate PETSc matrices in a simple
and safe fashion.

A design choice is that the building and allocation phase of a matrix is
completely separated from matrix operations. This means that once a matrix is
created it is structurally immutable. This restriction is slightly broken by +=,
*= and axpy operations, which possibly changes the underlying structure.
I'm not sure if I actually want to support structural mutability for matrices,
even through operators, so they might be removed in the future.

The supported operations are currently rather minimal. When a real-world
scenario requires some feature from the matrix it should be considered written
at that time.

The Buildinb phase is explicitly separated from the matrix phase through the
Matrix::Builder family of classes.