Introduces mixins for building PETSc classes
petsc uses opaque handles for their objects, on which we rely on unique_ptr to
manage lifetimes. However, in order to avoid repeating lifetime management code
across all different object times, a mixin is provided. This patch introduces
the mixin for the unique_ptr which provides the ::ptr() function, automatic
lifetime management, move constructor and operator= (which must be declared as
default in the derived class) and implicit conversion.

A benefit of this approach is that lifetime-related tests need only actually
test lifetimes, which means other tests now really won't have to. The mixin
completely separates resource responsibilty from business logic.