Refactors vector constructors to delegate
With the addition of petsc mixins, it is apparent that the previous scheme of
ad-hoc constructor delegation is no good, as the handle is no longer exposed.
Instead, this feature is now provided through a couple of free functions.
Nothing changes in terms of semantics, and the code is still exception safe
through the use of std::unique_ptr.

With full C++11 support this can be rewritten to use delegating constructors,
which should clean everything up nicely.